refactor(pricing): split Plans into per-card components
Plans was 308 lines, over the react-doctor no-giant-component gate.
Each card moves to its own file with the shared primitives in shared.js;
Plans.js keeps state and composition. CURRENCIES/formatPrice re-exported
so src/pages/pricing.js is untouched.
